+++ app/src/main/res/layout/fragment_parking_lot_info.xml
@@ -31,7 +31,31 @@
 
         app:textSize="16dp"/>
 
+    <ImageView
+        android:id="@+id/parkingLotImgView"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginEnd="16dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/terminal_segment_control"
+        tools:src="@drawable/parking_sample"
+        android:adjustViewBounds="true"
+        android:scaleType="centerCrop"/>
 
+    <androidx.recyclerview.widget.RecyclerView
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginBottom="16dp"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="16dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/parkingLotImgView" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
